Fix context provider logic from app to main
