+++ src/containers/AuthPage/AuthPage.jsx
@@ -5,7 +5,7 @@ import { supabase } from '../../helpers/supabaseClient';
 import { USER_MESSAGES } from '../../strings';
 import { ThemeContext, UserContext } from '../../context/context';
 import { Button, UserDashboard, SignForm } from '../../components';
-
+import { getUserData } from '../../api/api';
 import styles from './AuthPage.module.css';
 import cx from 'classnames';
