+++ src/containers/InstrumentsCatalogue/InstrumentsCatalogue.module.css
@@ -4,7 +4,7 @@
 .container {
   max-width: 1440px;
   margin: 0 auto;
-  padding: 25px;
+  margin-top: 30px;
   align-items: flex-start;
   width: 100%;
   display: flex;
