+++ src/context/context.jsx
@@ -1,7 +1,18 @@
 import { createContext } from 'react';
+import { useState } from 'react';
 
-export const UserContext = createContext(null);
+export const UserContext = createContext();
 
-export const ThemeContext = createContext(null);
+export const UserProvider = ({ children }) => {
+  const [user, setUser] = useState(null);
 
-export const SessionContext = createContext(false);
+  return <UserContext.Provider value={{ user, setUser }}>{children}</UserContext.Provider>;
+};
+
+export const ThemeContext = createContext('light');
+
+export const ThemeProvider = ({ children }) => {
+  const [theme, setTheme] = useState('light');
+
+  return <ThemeContext.Provider value={{ theme, setTheme }}>{children}</ThemeContext.Provider>;
+};
